[Refactor]: Remove chart from tokenWithChart
