--- cardstack/src/screens/sheets/TokenWithChartSheet/types.ts
@@ -1,6 +1,0 @@
-import { AssetWithNativeType } from '@cardstack/types';
-
-export interface AmountRowProps {
-  title: string;
-  asset: AssetWithNativeType;
-}
