+++ src/components/coin-row/BalanceCoinRowWrapper.tsx
@@ -39,7 +39,7 @@ const BalanceCoinWrapper = (item: AssetWithNativeType) => {
         select(item.address);
       }
     } else {
-      navigate(Routes.TOKEN_WITH_CHART_SHEET, { asset: item });
+      navigate(Routes.TOKEN_SHEET, { asset: item });
     }
   };
